fix: movement rounded and tied to delta t

+++ src/physics.rs
@@ -16,9 +16,9 @@ pub struct Velocity {
 
 #[derive(Component)]
 pub struct Position {
-    pub x: i32,
-    pub y: i32,
-    pub z: i32,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 // fn apply_gravity(
